# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "hamzaouhamou@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "signUp",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "email",
                                    "password",
                                    "last_name"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "last_name": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully"
                    },
                    "200": {
                        "description": "Register Successfully"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/user-authors": {
            "get": {
                "tags": [
                    "Author"
                ],
                "summary": "list all user-author records",
                "operationId": "d0245e41ef72a1ff8544fae75d2f50a0",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Author"
                ],
                "summary": "Add user-author",
                "operationId": "7f11bad2d6b17eb48f3e5b837ffc3937",
                "requestBody": {
                    "description": "user-author object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserAuthorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user-authors/{id}": {
            "delete": {
                "tags": [
                    "Author"
                ],
                "summary": "delete user-authors record",
                "operationId": "1d269d958e83664d07765dd4941cd16e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user-author id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "user-author object",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user-categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "list all user-category records",
                "operationId": "1e7524eb5784f5c5d76ca9f6cf569299",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Add user-category",
                "operationId": "c9c3c94ba01a456f38a4b2319ab820e2",
                "requestBody": {
                    "description": "user-category object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user-categories/{id}": {
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "delete user-category record",
                "operationId": "872ad02dd1601f78991800cd554acc7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user-categories id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "user-categories object",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserAuthorRequest": {
                "title": "User Author Request",
                "required": [
                    "user_id",
                    "author_name",
                    "author_id"
                ],
                "properties": {
                    "user_id": {
                        "description": "Class UserAuthor",
                        "type": "integer"
                    },
                    "author_name": {
                        "type": "string"
                    },
                    "author_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserCategoryRequest": {
                "title": "User Category Request",
                "required": [
                    "user_id",
                    "category_name",
                    "category_id"
                ],
                "properties": {
                    "user_id": {
                        "description": "Class UserCategory",
                        "type": "integer"
                    },
                    "category_name": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}